Flask-Backend/requirements.txt
Flask==3.1.0
Flask_SocketIO==5.4.1
matplotlib==3.7.4
mediapipe==0.10.15
numpy==1.26.4
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
Pillow==9.5.0
scikit_learn==1.2.2
gdown
